add double to printf

+++ ft_math_power.c
@@ -6,11 +6,11 @@
 /*   By: gapoulai <gapoulai@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/05 18:37:43 by gapoulai          #+#    #+#             */
-/*   Updated: 2020/12/05 18:39:31 by gapoulai         ###   ########lyon.fr   */
+/*   Updated: 2021/01/13 01:20:17 by gapoulai         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_math_power(int nb, int power)
+long long		ft_math_power(long long nb, long long power)
 {
 	if (power < 0)
 		return (0);
